requirements.txt
# Dependencias compartidas base
python-dotenv==1.0.0

